[Finishes #161499257] Update folder structure and add registration endpopoint
