+++ .gitignore
@@ -1,3 +1,4 @@
+.env
 *__pycache__
 *.cache
 *.pytest_cache
