+++ .travis.yml
@@ -1,0 +1,13 @@
+language: python
+python:
+  "3.6"
+install:
+  pip install -r requirements.txt
+env:
+    SECRET_KEY="\xa5%t\x|"
+before_script:
+ - psql -c ‘create database weconnect_test;’ -U postgres
+script:
+  pytest --cov=app
+after_success:
+  coveralls
