+++ app/v2/__init__.py
@@ -3,9 +3,9 @@ from flask import Blueprint
 from flask_restful import Api
 
 
-from app.api.v2.resources.products_endpoints import ProductEndpoint
+from app.v2.resources.products_endpoints import ProductEndpoint
 
-from app.api.v2.resources.auth_endpoints import UserRegistrationEndpoint
+from app.v2.resources.auth_endpoints import UserRegistrationEndpoint
 
 
 VERSION_2 = Blueprint('API2', __name__, url_prefix="/api/v2")
